% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_pair_bsample}
\alias{example_pair_bsample}
\title{A simulated perfectly correlated peak-gene pair with multiple biological samples present}
\format{
\subsection{\code{example_pair_bsample}}{

Three lists:
The first two list are named 'gene' and 'peak' respectively, each with 2 elements:
\describe{
\item{counts}{A length 2000 vector of gene/peak counts}
\item{seq_depths}{A length 2000 vector of sequencing depths of scRNA-seq/scATAC-seq}
},
the third list is named 'bsample' and hosts a length n vector of sample labels
}
}
\usage{
example_pair_bsample
}
\description{
This data set is created only for illustrative purposes and is used to test
scMultiMap_IRLS() and scMultiMap_WLS().
The source code for generating this data set is in data-raw/ on Github.
}
\keyword{datasets}
