% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_top_peak_gene_pairs}
\alias{get_top_peak_gene_pairs}
\title{get_top_peak_gene_pairs}
\usage{
get_top_peak_gene_pairs(
  obj,
  gene_top = 2000,
  peak_top = 20000,
  distance = 5e+05,
  gene_assay = "RNA",
  peak_assay = "peak"
)
}
\arguments{
\item{obj}{A Seurat object of single-cell multimodal data.}

\item{gene_top}{Number of top highly expressed genes. Default to 2000.}

\item{peak_top}{Number of top highly accessible peaks. Default to 20000.}

\item{distance}{Distance threshold for peaks to be considered as possibly associated with genes. Default to 5e+05.}

\item{gene_assay}{Name of the gene assay in Seurat object. Default to 'RNA'.}

\item{peak_assay}{Name of the peak assay in Seurat object. Default to 'peak'.}
}
\value{
A data frame of candidate peak-gene pairs.
\describe{
\item{gene}{gene in the peak-gene pair}
\item{peak}{peak in the peak-gene pair}
}
}
\description{
For peaks and genes with high abundance, construct candidate pairs where
peaks are in close proximity to the gene.
}
