% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMutliMap.R
\name{scMultiMap}
\alias{scMultiMap}
\title{scMultiMap}
\source{
Cell-type-specific mapping of enhancer and target genes from single-cell multimodal data.
Chang Su, Dongsoo Lee, Peng Jin, Jingfei Zhang;
https://www.biorxiv.org/content/10.1101/2024.09.24.614814v1
}
\usage{
scMultiMap(
  obj,
  pairs_df,
  bsample = NULL,
  gene_assay = "RNA",
  peak_assay = "peak",
  seq_depth_list = NULL,
  irls = T,
  verbose = F
)
}
\arguments{
\item{obj}{A Seurat object of single-cell multimodal data.}

\item{pairs_df}{A data frame of peak-gene pairs to evaluate. The first column is gene name,
and the second column is peak name.}

\item{bsample}{Default to NULL when all cells are from the same biological sample.
When multiple biological samples are present, \emph{bsample} should be either the column name for
the sample label in Seurat object meta data, or a length n vector of labels for biological samples.}

\item{gene_assay}{Name of the gene assay in Seurat object. Default to 'RNA'.}

\item{peak_assay}{Name of the peak assay in Seurat object. Default to 'peak'.}

\item{seq_depth_list}{A list of two length n vectors of sequencing depths.
If NULL, they will be calculated as the total number of gene/peak counts in each cell
for two modalities, respectively.}

\item{irls}{Whether to run IRLS to estimate mean and variance parameters in step 1.
Default to TRUE.}

\item{verbose}{Whether to print detailed messages. Default to FALSE.}
}
\value{
A data frame of scMultiMap results with five columns.
\describe{
\item{gene}{gene in the peak-gene pair}
\item{peak}{peak in the peak-gene pair}
\item{pval}{p-value}
\item{test_stat}{test statistics}
\item{covar}{estimated covariance}
}
Each row corresponds to one peak-gene pair's results from scMultiMap.
\emph{pval} and \emph{test_stat} denote the statistical significance of peak-gene association,
and \emph{covar} denote the magnitude of association.
The number of rows is equal to the number of pairs in \emph{pairs_df} present in the Seurat object.
}
\description{
Run scMultiMap to infer cell-type-specific peak-gene associations.
}
\examples{
# For illustrative purposes, we provide
# a toy Seurat object `small_obj` and
# a data frame of candidate peak-gene pairs `small_pairs_df`
# in this R package.
res <- scMultiMap(small_obj, small_pairs_df)
head(res,2)
#    gene                    peak      pval test_stat        covar
# 1 MALAT1 chr11-65013281-65014841 0.1521555 1.4319591 3.100282e-05
# 2 MALAT1 chr11-65026934-65027962 0.7938414 0.2613256 5.204042e-06

}
