% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMultiMap_IRLS.R
\name{scMultiMap_IRLS}
\alias{scMultiMap_IRLS}
\title{Iteratively reweighted least squares (IRLS) in scMultiMap}
\source{
Cell-type-specific mapping of enhancer and target genes from single-cell multimodal data.
Chang Su, Dongsoo Lee, Peng Jin, Jingfei Zhang;
https://www.biorxiv.org/content/10.1101/2024.09.24.614814v1
}
\usage{
scMultiMap_IRLS(X, seq_depth, bsample = NULL, irls = T, verbose = F)
}
\arguments{
\item{X}{A n by p count matrix, where n denotes the number of cells and p denotes the number of genes/peaks.
It can be a length p vector when n=1.}

\item{seq_depth}{A length n vector of sequencing depths}

\item{bsample}{A length n vector of the names of biological sample the cell is from. When all cells are from the same
sample, set to NULL (Default).}

\item{irls}{Whether to use IRLS. Default to TRUE. If FALSE, use ordinary least squares.}

\item{verbose}{Whether to print detailed messages. Default to FALSE.}
}
\value{
A list of two length p vectors:
\describe{
\item{mu}{estimated mean}
\item{sigma_sq}{estimated variance}
}
}
\description{
This function implements the 1st step of scMultiMap:
estimate mean and variance for peaks or genes with IRLS.
The output will be used as input to the 2nd step, scMultiMap_WLS(), to
estimate and infer the association between genes and peaks.
}
\examples{
# The following codes illustrate the use of `scMultiMap_IRLS` with toy datasets
# `small_obj` provided in this R package.
irls_list <- list()
irls_list[['gene']] <- scMultiMap_IRLS(Matrix::t(small_obj[['RNA']]$counts),
 Matrix::colSums(small_obj[['RNA']]$counts))
irls_list[['peak']] <- scMultiMap_IRLS(Matrix::t(small_obj[['peak']]$counts),
 Matrix::colSums(small_obj[['peak']]$counts))
print(str(irls_list[['gene']])) # estimated mean and variance for genes
print(str(irls_list[['peak']])) # estimated mean and variance for peaks

}
