% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMultiMap_WLS.R
\name{scMultiMap_WLS}
\alias{scMultiMap_WLS}
\title{Weighted least squares (WLS) in scMultiMap}
\source{
Cell-type-specific mapping of enhancer and target genes from single-cell multimodal data.
Chang Su, Dongsoo Lee, Peng Jin, Jingfei Zhang;
https://www.biorxiv.org/content/10.1101/2024.09.24.614814v1
}
\usage{
scMultiMap_WLS(count_list, seq_depth_list, pairs_df, irls_list, verbose = F)
}
\arguments{
\item{count_list}{A list of two p by n count matrices.
The first element is for gene and the second for peak.}

\item{seq_depth_list}{A list of two length p vectors of sequencing depths.
The first element is for scRNA-seq and the second for scATAC-seq.}

\item{pairs_df}{A data frame of candidate peak-gene pairs to evaluate.
The first column is gene name, and the second is peak name.}

\item{irls_list}{A list of two lists generated from the 1st step of scMultiMap,
i.e. \emph{scMultiMap_IRLS}. The first list contains the results on genes,
the second on peaks.}

\item{verbose}{Whether to print detailed messages. Default to FALSE.}
}
\value{
A data.frame with five columns
\describe{
\item{gene}{gene in the peak-gene pair}
\item{peak}{peak in the peak-gene pair}
\item{pval}{p-value}
\item{test_stat}{test statistics}
\item{covar}{estimated covariance}
}
Each row corresponds to one peak-gene pair's results from scMultiMap.
\emph{pval} and \emph{test_stat} denote the statistical significance of peak-gene association,
and \emph{covar} denote the magnitude of association.
The number of rows is equal to the number of pairs in \emph{pairs_df} present in the Seurat object.
}
\description{
This function implements the 2nd step of scMultiMap:
estimate and infer peak-gene association with WLS.
It takes the output from the 1st step, \emph{scMultiMap_IRLS} as input.
}
\examples{
count_list <- list(gene = Matrix::t(small_obj[['RNA']]$counts),
 peak = Matrix::t(small_obj[['peak']]$counts))
seq_depth_list <- list(gene = Matrix::colSums(small_obj[['RNA']]$counts),
 peak = Matrix::colSums(small_obj[['peak']]$counts))
wls_res <- scMultiMap_WLS(count_list, seq_depth_list, small_pairs_df, small_irls_list)

}
