% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMultiMap_WLS_by_gene.R
\name{scMultiMap_WLS_by_gene}
\alias{scMultiMap_WLS_by_gene}
\title{WLS utility function.}
\source{
Cell-type-specific mapping of enhancer and target genes from single-cell multimodal data.
Chang Su, Dongsoo Lee, Peng Jin, Jingfei Zhang;
}
\usage{
scMultiMap_WLS_by_gene(
  x_gene,
  X_peak,
  m_gene,
  M_peak,
  w_gene,
  W_peak,
  var_gene,
  Var_peak,
  gene_seq_depth,
  peak_seq_depth
)
}
\arguments{
\item{x_gene}{A length n vector of gene counts. n denotes the number of cells.}

\item{X_peak}{A n by p peak count matrix. p denotes the number of peaks.}

\item{m_gene}{A length n vector of gene counts' mean across cells.}

\item{M_peak}{A n by p matrix of peak counts' mean across cells for all peaks.}

\item{w_gene}{A length n vector used to calculate regression weights.}

\item{W_peak}{A n by p matrix used to calculate regression weights.}

\item{var_gene}{A length n vector of gene counts' variance across cells.}

\item{Var_peak}{A n by p matrix of peak counts' variance across cells for all peaks.}

\item{gene_seq_depth}{A length n vector of sequencing depths from the RNA modality.}

\item{peak_seq_depth}{A length n vector of sequencing depths from the ATAC/peak modality.}
}
\value{
A data.frame with p rows and three columns
\describe{
\item{pval}{p-value}
\item{test_stat}{test statistics}
\item{covar}{estimated covariance}
}
Each row correspond to scMultiMap's WLS result on one gene-peak pair.
}
\description{
This function performs WLS for one gene and all peaks in its candidate pairs.
}
