% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_irls_list}
\alias{small_irls_list}
\title{A toy example of estimated results from scMultiMap_IRLS() on \code{small_obj}.}
\format{
\subsection{\code{small_irls_list}}{

A list of two lists generated from the 1st step of scMultiMap
\describe{
\item{gene}{\emph{scMultiMap_IRLS} results on genes}
\item{peak}{\emph{scMultiMap_IRLS} results on peaks}
}
}
}
\usage{
small_irls_list
}
\description{
The source code for generating this data set is in data-raw/ on Github.
}
\keyword{datasets}
