% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_obj}
\alias{small_obj}
\title{A small pedagogical Seurat object}
\format{
\subsection{\code{small_obj}}{

A Seurat object with 100 cells and two assay
\describe{
\item{RNA}{A gene count matrix with 100 genes and 100 cells}
\item{peak}{A peak count matrix with 250 genes and 100 cells}
}
}
}
\usage{
small_obj
}
\description{
This Seurat object is created only for illustrative purposes and is used to test
scMultiMap().
The source code for generating this data set is in data-raw/ on Github.
}
\keyword{datasets}
