% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{small_pairs_df}
\alias{small_pairs_df}
\title{A data frame of peak-gene pairs, to be used with \code{small_obj}}
\format{
\subsection{\code{small_pairs_df}}{

A data frame with 250 rows (peak-gene pairs) and 2 columns
\describe{
\item{gene}{gene name}
\item{peak}{peak name}
}
}
}
\usage{
small_pairs_df
}
\description{
This data frame holds the candidate peak-gene pairs to be inferred with
scMultiMap() and \code{small_obj}.
The source code for generating this data set is in data-raw/ on Github.
}
\keyword{datasets}
